requirements.txt
# Core Python dependencies for Prajna Voice Assistant
numpy>=2.0.0
sounddevice>=0.4.6
webrtcvad-wheels==2.0.14
openwakeword==0.4.0
onnxruntime>=1.20.0
